Excecoes completas

+++ src/main/java/com/_clientes/demo/services/ClienteService.java
@@ -3,10 +3,15 @@ package com._clientes.demo.services;
 import com._clientes.demo.dto.ClienteDTO;
 import com._clientes.demo.entities.Cliente;
 import com._clientes.demo.repositories.ClienteRepository;
+import com._clientes.demo.services.exceptions.DatabaseException;
+import com._clientes.demo.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,8 +22,9 @@ public class ClienteService {
 
     @Transactional(readOnly = true)
     public ClienteDTO findById(Long id) {
-        Cliente cliente = repository.findById(id).get() ;
-        return new ClienteDTO(cliente);
+            Cliente cliente = repository.findById(id).orElseThrow(
+                    () -> new ResourceNotFoundException("Recurso não encontrado"));
+            return new ClienteDTO(cliente);
     }
 
     @Transactional(readOnly = true)
@@ -37,15 +43,29 @@ public class ClienteService {
 
     @Transactional
     public ClienteDTO update(Long id, ClienteDTO dto) {
-        Cliente entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClienteDTO(entity);
+        try {
+            Cliente entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClienteDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id); ;
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ClienteDTO dto, Cliente entity) {
